Extract control/sort kernels into kernels/ directory
Move kernel algorithm code (templates, functions, types, constants) from
test .cpp files into reusable header-only .hpp files under kernels/:

- kernels/control/hashtable_lookup_simd.hpp (SIMD tile-based lookup)
- kernels/control/hashtable_lookup_simt.hpp (SIMT thread-based lookup)
- kernels/sort/topk.hpp (SIMT per-bucket top-k)

Test .cpp files now contain only the benchmark harness: extern data
declarations, test-specific constants (kNum, kCapacity, etc.), global
buffers, and main(). Each includes its kernel header via the -I/kernels
include path (e.g. #include "control/hashtable_lookup_simd.hpp").

Also fix compile_all.sh to use local toolchain paths via env vars.
